requirements.txt
Django
gunicorn
psycopg2-binary
djangorestframework
django-rest-swagger
django-filter
django-cors-headers
djangorestframework-jwt
PyJWT
viberbot
Pillow